Split backtesting and optimization. Split optimization and result saving. Import only what is needed from passivbot.
